+++ src/Controller/CoursController.php
@@ -37,6 +37,10 @@ class CoursController extends AbstractController
     #[Route('/{id}/new', name: 'app_cours_new', methods: ['GET', 'POST'])]
     public function new(Niveau $niveau,Request $request, EntityManagerInterface $entityManager): Response
     {
+        if(!isset($usr->getRoles()["app_cours_new"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_home_access_denied', [], Response::HTTP_SEE_OTHER);
+        }
         $usr = $this->getUser();
         $sections = $entityManager
             ->getRepository(Section::class)
@@ -83,6 +87,10 @@ class CoursController extends AbstractController
     #[Route('/{id}/edit', name: 'app_cours_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Cours $cour, EntityManagerInterface $entityManager): Response
     {
+        if(!isset($usr->getRoles()["app_cours_edit"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_home_access_denied', [], Response::HTTP_SEE_OTHER);
+        }
         $usr = $this->getUser();
         $sections = $entityManager
             ->getRepository(Section::class)
@@ -110,7 +118,10 @@ class CoursController extends AbstractController
     public function delete(Request $request, Cours $cour, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
-        
+        if(!isset($usr->getRoles()["app_cours_delete"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_home_access_denied', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$cour->getId(), $request->request->get('_token'))) {
             $entityManager->remove($cour);
             $entityManager->flush();
